refactor(editors/substation/sub-function-editor): rename custom element
